Tidy up the editing and display of items
