+++ module/data/melee-data.mjs
@@ -5,7 +5,7 @@ export class MeleeDataModel extends foundry.abstract.TypeDataModel {
     const { fields } = foundry.data;
     return {
       ...dlcFields.setting(),
-      db: new fields.StringField({ required: true, blank: false }),
+      ...dlcFields.integerNoMax('db', 0), // Defence bonus
       notes: new fields.HTMLField({ required: false }),
       ...dlcFields.integerNoMax('price', 0), // in cents
       damage: new fields.StringField({
