+++ module/init/register-item-sheets.mjs
@@ -1,8 +1,8 @@
-import { ShowEdgeSheet } from '../sheets/show-edge-sheet.mjs';
-import { ShowGunSheet } from '../sheets/show-gun-sheet.mjs';
-import { ShowMeleeSheet } from '../sheets/show-melee-sheet.mjs';
-import { ShowMiscItemSheet } from '../sheets/show-misc-item-sheet.mjs';
-import { ShowOtherRangedSheet } from '../sheets/show-other-ranged-sheet.mjs';
+import { ShowEdgeSheet } from '../sheets/show-sheet-edge.mjs';
+import { ShowGunSheet } from '../sheets/show-sheet-gun.mjs';
+import { ShowMeleeSheet } from '../sheets/show-sheet-melee.mjs';
+import { ShowMiscItemSheet } from '../sheets/show-sheet-misc-item.mjs';
+import { ShowOtherRangedSheet } from '../sheets/show-sheet-other-ranged.mjs';
 
 export function registerItemSheets() {
   Items.unregisterSheet('core', ItemSheet);
