+++ module/sheets/edit-sheet-edge.mjs
@@ -1,42 +1,19 @@
-export class ShowEdgeSheet extends ItemSheet {
+import { ItemSheetBase } from './item-sheet-base.mjs';
+
+export class EditEdgeSheet extends ItemSheetBase {
   constructor(data, options = {}) {
-    // setting up tabs here instead of in defaultOptions so that we can easily set the initial tab
-    // eslint-disable-next-line no-param-reassign
-    options.tabs = [
-      {
-        navSelector: '.sheet-tabs',
-        contentSelector: '.sheet-body',
-        initial: 'configure',
-      },
-    ];
     super(data, options);
   }
 
   static get defaultOptions() {
     return foundry.utils.mergeObject(super.defaultOptions, {
-      classes: ['dlc', 'sheet', 'item'],
-      template: 'systems/deadlands-classic/templates/item/show-edge-sheet.html',
-      width: 660,
-      height: 800,
-      closeOnSubmit: false,
-      submitOnClose: false,
-      submitOnChange: true,
+      template: 'systems/deadlands-classic/templates/item/edit-edge-sheet.html',
     });
   }
 
   /** @override */
-  async getData() {
-    const context = {
-      cssClass: 'locked',
-      editable: false,
-      limited: this.document.limited,
-      name: this.title,
-      options: this.options,
-      owner: this.document.isOwner,
-      system: this.item.system,
-      title: this.title,
-      img: this.document.img,
-    };
+  async getData(options) {
+    const context = await super.getData(options);
 
     context.enrichedBlurb = await TextEditor.enrichHTML(
       this.item.system.blurb,
