+++ module/sheets/show-sheet-edge.mjs
@@ -1,44 +1,24 @@
-export class EditEdgeSheet extends ItemSheet {
+import { ItemSheetBase } from './item-sheet-base.mjs';
+
+export class ShowEdgeSheet extends ItemSheetBase {
   constructor(data, options = {}) {
-    // setting up tabs here instead of in defaultOptions so that we can easily set the initial tab
-    // eslint-disable-next-line no-param-reassign
-    options.tabs = [
-      {
-        navSelector: '.sheet-tabs',
-        contentSelector: '.sheet-body',
-        initial: 'configure',
-      },
-    ];
     super(data, options);
   }
 
   static get defaultOptions() {
     return foundry.utils.mergeObject(super.defaultOptions, {
-      classes: ['dlc', 'sheet', 'item'],
-      template: 'systems/deadlands-classic/templates/item/edit-edge-sheet.html',
-      width: 660,
-      height: 800,
-      closeOnSubmit: false,
-      submitOnClose: false,
-      submitOnChange: true,
+      template: 'systems/deadlands-classic/templates/item/show-edge-sheet.html',
     });
   }
 
   /** @override */
   async getData(options) {
-    const { isEditable } = this;
+    let context = await super.getData(options);
 
-    const context = {
-      cssClass: isEditable ? 'editable' : 'locked',
-      editable: isEditable,
-      limited: this.document.limited,
-      name: this.title,
-      options: this.options,
-      owner: this.document.isOwner,
-      system: this.item.system,
-      title: this.title,
-      img: this.document.img,
-    };
+    context = foundry.utils.mergeObject(context, {
+      cssClass: 'locked',
+      editable: false,
+    });
 
     context.enrichedBlurb = await TextEditor.enrichHTML(
       this.item.system.blurb,
