+++ module/sheets/show-sheet-other-ranged.mjs
@@ -1,0 +1,26 @@
+import { ItemSheetBase } from './item-sheet-base.mjs';
+
+export class ShowOtherRangedSheet extends ItemSheetBase {
+  constructor(data, options = {}) {
+    super(data, options);
+  }
+
+  static get defaultOptions() {
+    return foundry.utils.mergeObject(super.defaultOptions, {
+      template:
+        'systems/deadlands-classic/templates/item/show-other-ranged-sheet.html',
+    });
+  }
+
+  /** @override */
+  async getData(options) {
+    let context = await super.getData(options);
+
+    context = foundry.utils.mergeObject(context, {
+      cssClass: 'locked',
+      editable: false,
+    });
+
+    return context;
+  }
+}
