+++ templates/item/edit-melee-sheet.html
@@ -1,6 +1,6 @@
 <form class="{{cssClass}}" autocomplete="off">
-    <header class="sheet-header">
-        <img src="{{item.img}}" data-edit="img" title="{{item.name}}" height="64" width="64"/>
-        <h1><input name="name" type="text" value="{{item.name}}" placeholder="{{ localize 'Name' }}"/></h1>
-    </header>
+  <header class="sheet-header">
+    <img src={{img}} data-edit="img" title={{img}} height="64" width="64"/>
+    <h1><input name="name" type="text" value="{{this.title}}" placeholder="{{this.title}}"/></h1>
+  </header>
 </form>
